EAGLE-85 : Fix unit test failure.
