+++ eagle-external/eagle-log4jkafka/src/test/scala/org/apache/eagle/log4j/kafka/TestGenericLogKeyer.scala
@@ -22,27 +22,25 @@ import java.util.Properties
 import org.apache.eagle.log4j.kafka.hadoop.GenericLogKeyer
 import org.scalatest.{FlatSpec, Matchers}
 
-
-class TestGenericLogKeyer extends FlatSpec with Matchers  {
-
-  val hdfsMsg = "2015-07-31 01:54:35,161 INFO FSNamesystem.audit: allowed=true ugi=root (auth:TOKEN) ip=/10.0.0.1 cmd=open src=/tmp/private dst=null perm=null"
-  val props = new Properties()
-  props.put("keyPattern", "ugi=(\\w+)[@\\s+]")
-  props.put("keyPattern2", "user=(\\w+),\\s+")
-  val test = new GenericLogKeyer(props)
-  var keyVal = test.getKey(hdfsMsg)
-  println(keyVal)
-
-  val hbaseMsg = "2015-11-06 13:14:00,741 TRACE SecurityLogger.org.apache.hadoop.hbase.security.access.AccessController: Access allowed for user root; reason: All users allowed; remote address: /192.168.56.101; request: scan; context: (user=root, scope=hbase:meta, family=info, action=READ)"
-  props.put("keyPattern", "user=(\\w+),\\s+")
-  keyVal = test.getKey(hbaseMsg)
-  println(keyVal)
-
-  //props.put("keyPattern", "user=(\\w+),\\s+")
-  val props1 = new Properties()
-  val test1 = new GenericLogKeyer(props1)
-  keyVal = test1.getKey(hbaseMsg)
-  println(keyVal)
-
+abstract class TestGenericLogKeyer extends FlatSpec with Matchers  {
+  
+    val hdfsMsg = "2015-07-31 01:54:35,161 INFO FSNamesystem.audit: allowed=true ugi=root (auth:TOKEN) ip=/10.0.0.1 cmd=open src=/tmp/private dst=null perm=null"
+    val props = new Properties()
+    props.put("keyPattern", "ugi=(\\w+)[@\\s+]")
+    props.put("keyPattern2", "user=(\\w+),\\s+")
+    val test = new GenericLogKeyer(props)
+    var keyVal = test.getKey(hdfsMsg)
+    println(keyVal)
+  
+    val hbaseMsg = "2015-11-06 13:14:00,741 TRACE SecurityLogger.org.apache.hadoop.hbase.security.access.AccessController: Access allowed for user root; reason: All users allowed; remote address: /192.168.56.101; request: scan; context: (user=root, scope=hbase:meta, family=info, action=READ)"
+    props.put("keyPattern", "user=(\\w+),\\s+")
+    keyVal = test.getKey(hbaseMsg)
+    println(keyVal)
+  
+    //props.put("keyPattern", "user=(\\w+),\\s+")
+    val props1 = new Properties()
+    val test1 = new GenericLogKeyer(props1)
+    keyVal = test1.getKey(hbaseMsg)
+    println(keyVal)
 
 }
